# docs/swagger.yaml
basePath: /
definitions:
...
  handler.ColumnResponse:
    properties:
      board_id:
        type: string
      id:
        type: string
      position:
        type: integer
      title:
        type: string
    type: object
...
  handler.CreateColumnRequest:
    properties:
      board_id:
        type: string
      position:
        type: integer
      title:
        type: string
    required:
    - board_id
    - title
    type: object
...
  handler.ReorderColumnsRequest:
    properties:
      columns:
        items:
          properties:
            id:
              type: string
            position:
              type: integer
          required:
          - id
          - position
          type: object
        type: array
    required:
    - columns
    type: object
...
  handler.UpdateColumnRequest:
    properties:
      position:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: you@example.com
    name: Your Name
    url: http://your-website.com
  description: API for managing Kanban boards.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Kanban API
  version: "1.0"
paths:
...
  /boards/{id}/columns:
    get:
      consumes:
      - application/json
      description: Retrieves all columns for the specified board, sorted by position
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Board ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board columns
          schema:
            items:
              $ref: '#/definitions/handler.ColumnResponse'
            type: array
        "400":
          description: Invalid board ID
          schema:
            type: object
        "401":
          description: Not authenticated
          schema:
            type: object
        "403":
          description: Insufficient permissions
          schema:
            type: object
        "500":
          description: Server error
          schema:
            type: object
      summary: Get all columns for a board
      tags:
      - Columns
  /boards/{id}/columns/reorder:
    post:
      consumes:
      - application/json
      description: Changes the order of columns on a board
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Board ID
        in: path
        name: id
        required: true
        type: string
      - description: Column reordering data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ReorderColumnsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            type: object
        "400":
          description: Invalid request data
          schema:
            type: object
        "401":
          description: Not authenticated
          schema:
            type: object
        "403":
          description: Insufficient permissions
          schema:
            type: object
        "500":
          description: Server error
          schema:
            type: object
      summary: Reorder board columns
      tags:
      - Columns
...
  /columns:
    post:
      consumes:
      - application/json
      description: Creates a new column on a board
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Column creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateColumnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created column
          schema:
            $ref: '#/definitions/handler.ColumnResponse'
        "400":
          description: Invalid request data
          schema:
            type: object
        "401":
          description: Not authenticated
          schema:
            type: object
        "403":
          description: Insufficient permissions
          schema:
            type: object
        "500":
          description: Server error
          schema:
            type: object
      summary: Create a new column
      tags:
      - Columns
  /columns/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a column by its ID
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Column ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            type: object
        "400":
          description: Invalid column ID
          schema:
            type: object
        "401":
          description: Not authenticated
          schema:
            type: object
        "403":
          description: Insufficient permissions
          schema:
            type: object
        "404":
          description: Column not found
          schema:
            type: object
        "500":
          description: Server error
          schema:
            type: object
      summary: Delete a column
      tags:
      - Columns
    get:
      consumes:
      - application/json
      description: Retrieves a column by its ID
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Column ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Column data
          schema:
            $ref: '#/definitions/handler.ColumnResponse'
        "400":
          description: Invalid column ID
          schema:
            type: object
        "401":
          description: Not authenticated
          schema:
            type: object
        "403":
          description: Insufficient permissions
          schema:
            type: object
        "404":
          description: Column not found
          schema:
            type: object
        "500":
          description: Server error
          schema:
            type: object
      summary: Get column by ID
      tags:
      - Columns
    put:
      consumes:
      - application/json
      description: Updates a column's details
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Column ID
        in: path
        name: id
        required: true
        type: string
      - description: Column update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateColumnRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated column
          schema:
            $ref: '#/definitions/handler.ColumnResponse'
        "400":
          description: Invalid request data
          schema:
            type: object
        "401":
          description: Not authenticated
          schema:
            type: object
        "403":
          description: Insufficient permissions
          schema:
            type: object
        "404":
          description: Column not found
          schema:
            type: object
        "500":
          description: Server error
          schema:
            type: object
      summary: Update a column
      tags:
      - Columns
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"